agents/requirements.txt
a2a==0.44
a2a-sdk==0.3.22
ag-ui-adk==0.4.2
ag-ui-protocol==0.1.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
Authlib==1.6.6
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.4
cssselect==1.4.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
fastapi==0.123.10
fastuuid==0.14.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
google-adk==1.21.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.136.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.0
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.8
langchain-core==1.2.8
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
litellm==1.81.8
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.26.0
mmh3==5.2.0
multidict==6.7.1
openai==2.16.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
parsel==1.11.0
propcache==0.4.1
Protego==0.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
PyJWT==2.11.0
pymongo==4.16.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
queuelib==1.9.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
Scrapy==2.14.1
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.3
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.22.2
tqdm==4.67.3
Twisted==25.5.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
w3lib==2.4.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.2
zstandard==0.25.0
